// server/src/modules/settings/types.ts
export interface UpdateCompanyInput {
  companyName?: string;
  contactPerson?: string;
  mobile?: string;
  email?: string;
  address?: string;
  gstNumber?: string;
  website?: string;
  bankName?: string;
  bankAccountName?: string;
  bankAccountNumber?: string;
  bankBranch?: string;
  bankIfsc?: string;
  bankUpi?: string;
  authorizedSignatory?: string;
  footerMessage?: string;
  termsAndConditions?: string;
}